+++ python/ray/train/tests/test_huggingface_predictor.py
@@ -26,6 +26,14 @@ tokenizer_checkpoint = "sgugger/gpt2-like-tokenizer"
 
 
 @pytest.fixture
+def ray_start_4_cpus():
+    address_info = ray.init(num_cpus=4)
+    yield address_info
+    # The code after the yield will run as teardown code.
+    ray.shutdown()
+
+
+@pytest.fixture
 def ray_start_runtime_env():
     # Requires at least torch 1.11 to pass
     # TODO update torch version in requirements instead
