+++ python/ray/train/tests/test_sklearn_predictor.py
@@ -11,7 +11,6 @@ from sklearn.ensemble import RandomForestClassifier
 
 import ray
 import ray.cloudpickle as cpickle
-from ray.air._internal.checkpointing import save_preprocessor_to_dir
 from ray.air.checkpoint import Checkpoint
 from ray.air.constants import MODEL_KEY
 from ray.data.preprocessor import Preprocessor
